+++ src/lib/converter/converter.ts
@@ -94,7 +94,8 @@ export class Converter extends ChildableComponent<
 
     /**
      * Triggered when the converter has created a signature reflection.
-     * The listener will be given {@link Context}, {@link SignatureReflection} | {@link ProjectReflection} and a `ts.Node?`
+     * The listener will be given {@link Context}, {@link SignatureReflection} | {@link ProjectReflection} and
+     * `ts.SignatureDeclaration | ts.IndexSignatureDeclaration | ts.JSDocSignature | undefined`
      * @event
      */
     static readonly EVENT_CREATE_SIGNATURE = ConverterEvents.CREATE_SIGNATURE;
@@ -184,6 +185,18 @@ export class Converter extends ChildableComponent<
     }
 
     /**
+     * Parse the given file into a comment. Intended to be used with markdown files.
+     */
+    parseRawComment(file: MinimalSourceFile) {
+        return parseComment(
+            lexCommentString(file.text),
+            this.config,
+            file,
+            this.application.logger
+        );
+    }
+
+    /**
      * Compile the files within the given context and convert the compiler symbols to reflections.
      *
      * @param context  The context object describing the current state the converter is in.
@@ -227,15 +240,6 @@ export class Converter extends ChildableComponent<
         const symbol = getSymbolForModuleLike(context, node);
         let moduleContext: Context;
 
-        const allExports = getExports(context, node, symbol);
-
-        if (allExports.every((exp) => this.shouldIgnore(exp))) {
-            this.owner.logger.verbose(
-                `All members of ${entryName} are excluded, ignoring entry point.`
-            );
-            return;
-        }
-
         if (singleEntryPoint) {
             // Special case for when we're giving a single entry point, we don't need to
             // create modules for each entry. Register the project as this module.
@@ -264,11 +268,8 @@ export class Converter extends ChildableComponent<
 
             if (entryPoint.readmeFile) {
                 const readme = readFile(entryPoint.readmeFile);
-                const comment = parseComment(
-                    lexCommentString(readme),
-                    context.converter.config,
-                    new MinimalSourceFile(readme, entryPoint.readmeFile),
-                    context.logger
+                const comment = this.parseRawComment(
+                    new MinimalSourceFile(readme, entryPoint.readmeFile)
                 );
 
                 if (comment.blockTags.length || comment.modifierTags.size) {
@@ -292,6 +293,7 @@ export class Converter extends ChildableComponent<
             moduleContext = context.withScope(reflection);
         }
 
+        const allExports = getExports(context, node, symbol);
         for (const exp of allExports.filter((exp) =>
             isDirectExport(context.resolveAliasedSymbol(exp), node)
         )) {