+++ src/test/converter2/issues/gh1980.ts
@@ -1,0 +1,6 @@
+/**
+ * {@link   http://example.com   }
+ * {@link   http://example.com  |   with text}
+ * {@link   http://example.com   jsdoc support}
+ */
+export const link = 123;