+++ src/test/converter2/issues/gh1996.ts
@@ -1,0 +1,3 @@
+export const a = () => {};
+
+export function b() {}