+++ src/test/converter2/issues/gh2008.ts
@@ -1,0 +1,4 @@
+const makeFn = () => () => {};
+
+/** Docs */
+export const myFn = makeFn();